fix array out of bounds in loadImages.js
Using Math.ceil(Math.random() * imgLen) can give an index 1 past the end of the array, which is likely to happen if there aren't very many items in imagesFileNames. 

The first item in imagesFileNames is usually going to be .DS_Store, though, which will lead to a segfault.

Simple workaround.